move loggers_v2 to logger for src and tests, pass logger tests
