+++ src/deepsparse/loggers/filters/pattern.py
