+++ src/deepsparse/loggers/registry/functions.py
