+++ src/deepsparse/loggers/registry/loggers/base_logger.py
