+++ src/deepsparse/loggers/utils.py
@@ -17,7 +17,7 @@ import re
 from typing import Any, Type
 
 
-LOGGER_REGISTRY = "src.deepsparse.loggers_v2.registry.__init__"
+LOGGER_REGISTRY = "src.deepsparse.loggers.registry.__init__"
 
 
 def import_from_registry(name: str) -> Type[Any]:
