+++ src/deepsparse/pipeline.py
@@ -17,7 +17,7 @@ import os
 from pathlib import Path
 from typing import Any, AsyncGenerator, Callable, Dict, Generator, List, Optional, Union
 
-from deepsparse.loggers_v2.logger_manager import LoggerManager
+from deepsparse.loggers.logger_manager import LoggerManager
 from deepsparse.middlewares import IS_NESTED_KEY, NAME_KEY, MiddlewareManager
 from deepsparse.operators import EngineOperator, Operator
 from deepsparse.pipeline_config import PipelineConfig
