+++ tests/deepsparse/loggers/registry/loggers/list_logger.py
