+++ tests/deepsparse/loggers/registry/loggers/test_prometheus_loggers.py
@@ -16,8 +16,8 @@ import requests
 
 import pytest
 from deepsparse.legacy.loggers import MetricCategories
-from deepsparse.loggers_v2.filters import unravel_value_as_generator
-from deepsparse.loggers_v2.registry.loggers.prometheus_logger import (
+from deepsparse.loggers.filters import unravel_value_as_generator
+from deepsparse.loggers.registry.loggers.prometheus_logger import (
     PrometheusLogger,
     get_prometheus_metric,
 )
